+++ src/v1/MediaType/matchesContentType.ts
@@ -32,11 +32,23 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./isMediaType";
-export * from "./matchesContentType";
-export * from "./mustBeMediaType";
-export * from "./mustMatchContentType";
-export * from "./MediaType";
-export * from "./MediaTypeParts";
-export * from "./parseContentType";
-export * from "./parseMediaType";
+import { MediaType } from "./MediaType";
+
+/**
+ * Data guard. Returns `true` if your `input` matches any of the MediaTypes
+ * in the `safelist`.
+ *
+ * We compare everything except the parameters of the MediaTypes.
+ *
+ * Use `mustMatchMediaType()` for the corresponding data guarantee.
+ */
+export function matchesContentType(input: MediaType, safelist: MediaType[]): boolean {
+    // what content type do we have?
+    const inputType = input.getContentType();
+
+    // we only need a single match
+    return safelist.some((possible) => {
+        const possibleType = possible.getContentType();
+        return inputType === possibleType;
+    });
+}